requirements.txt
geopandas==1.0.1
matplotlib==3.1.2
numpy==1.19.5
numpy==1.23.5
numpy==1.24.4
Pillow==11.2.1
protobuf==3.20.3
protobuf==3.19.6
protobuf==4.25.3
rasterio==1.4.2
rasterio==1.2.6
rasterio==1.4.3
Shapely==2.1.1
streamlit==1.37.1
tensorflow==2.5.1
tensorflow==2.10.1
tensorflow==2.18.0
tensorflow==2.19.0
tensorflow_gpu==2.3.2
tensorflow_intel==2.18.0
